# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dolar API & BCV Exchange Rates SaaS",
    "version": "1.0.0",
    "description": "API comercial de alta disponibilidad para consulta de tasas oficiales del Banco Central de Venezuela (USD, EUR, etc.), conversión de divisas a 2 decimales, gestión de suscripciones con Pago Móvil y Binance Pay, y administración de API Keys.",
    "contact": {
      "name": "Soporte Técnico",
      "email": "soporte@dolarapi.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3003",
      "description": "Servidor Local de Desarrollo"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Clave de API provista por la plataforma (ej. bcv_live_...)"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de sesión obtenido tras el login en el portal"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/user/keys": {
      "get": {
        "tags": ["API Keys"],
        "summary": "Listar API Keys del desarrollador",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de claves enmascaradas" }
        }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Crear nueva API Key",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Servidor Node Producción" },
                  "environment": { "type": "string", "enum": ["live", "test"], "example": "live" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Clave generada en texto plano una sola vez" }
        }
      }
    },
...
  }
}